test: move #33227 regression test to its own file
#19211 was fixed then regressed, so this is a true regression; isolate it
from the timing-sensitive streaming tests in fetch.stream.test.ts.
